recommender/requirements.txt
fastapi
uvicorn[standard]
pymongo
sentence-transformers
scikit-learn
python-dotenv
bson
